requirements.txt
oead
PySide2
numpy
pillow
PyGLM
PyOpenGL